refactor serialization: implement asdict(), from_dict()
